+++ activerecord/test/schema/schema.rb
@@ -106,6 +106,10 @@ ActiveRecord::Schema.define do
     t.column :font_size, :integer, default: 0
     t.column :difficulty, :integer, default: 0
     t.column :cover, :string, default: "hard"
+    t.string :isbn
+    t.datetime :published_on
+    t.index [:author_id, :name], unique: true
+    t.index :isbn, where: "published_on IS NOT NULL", unique: true
   end
 
   create_table :booleans, force: true do |t|